refactor: make getTempPassword a subcommand
