+++ src/main/java/org/justserve/cli/command/GetTempPassword.java
@@ -11,19 +11,16 @@ import org.justserve.model.UserHashRequestByEmail;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
 
-import java.awt.datatransfer.StringSelection;
-import java.util.concurrent.Callable;
-
-import static java.awt.Toolkit.getDefaultToolkit;
 import static org.justserve.cli.util.JustServePrinter.printError;
+import static org.justserve.cli.util.JustServePrinter.printNormal;
 
 @Command(name = "getTempPassword", description = "get a temporary password for a user")
-public class GetTempPassword extends BaseCommand implements Callable<Integer> {
+public class GetTempPassword extends BaseCommand implements Runnable {
 
     @ReflectiveAccess
     @Option(names = {"-e", "--email"}, description = "email for the user whose temporary password will be generated")
     String email;
-    
+
     @Inject
     @ReflectiveAccess
     Provider<UserClient> userClientProvider;
@@ -31,19 +28,14 @@ public class GetTempPassword extends BaseCommand implements Callable<Integer> {
     @Value("${justserve.token}")
     String token;
 
-    /**
-     * Computes a result, or throws an exception if unable to do so.
-     *
-     * @return computed result
-     */
     @Override
-    public Integer call() {
+    public void run() {
         HttpResponse<String> response;
         if ("i-need-to-be-defined".equals(token) || null == token) {
             printError(("NO AUTHENTICATION PROVIDED" + System.lineSeparator() +
                     "The Authentication token is not assigned as an environment variable." + System.lineSeparator() +
                     "Please define the environment variable \"JUSTSERVE_TOKEN\" and try again."));
-            return 1;
+            return;
         }
         try {
             UserClient userClient = userClientProvider.get();
@@ -52,15 +44,12 @@ public class GetTempPassword extends BaseCommand implements Callable<Integer> {
             String errorMessage = "Received an unexpected response from JustServe:" +
                     String.format("%n%d (%s)", e.getResponse().status().getCode(), e.reason());
             printError(errorMessage);
-            return 1;
+            return;
         }
         if (response != null) {
-            getDefaultToolkit().getSystemClipboard().setContents(
-                    new StringSelection(response.body().replace("\"", "").trim()), null);
-            return 0;
+            printNormal(response.body().replace("\"", "").trim());
         } else {
             printError("An unexpected error occurred. Response from JustServe was null.");
-            return 1;
         }
     }
 }
